+++ inst/doc/fluxible.R
@@ -11,24 +11,21 @@ library(fluxible)
 str(co2_liahovden, width = 70, strict.width = "cut")
 
 ## ----record_liahovden-str, message=FALSE, echo=FALSE--------------------------
-
 str(record_liahovden, width = 70, strict.width = "cut")
 
-
-## ----match, message=FALSE, echo=TRUE------------------------------------------
+## ----match, message=FALSE-----------------------------------------------------
 library(fluxible)
 
 conc_liahovden <- flux_match(
   raw_conc = co2_liahovden, # dataframe with raw gas concentration
   field_record = record_liahovden, # dataframe with meta data
-  f_datetime = datetime, # column containing date and time
+  f_datetime = datetime, # date and time of each gas concentration row
   start_col = start, # start date and time of each measurement
   measurement_length = 220, # length of measurements (in seconds)
-  fixed_length = TRUE, # the length of the measurement is a constant
   time_diff = 0 # time difference between f_datetime and start_col
 )
 
-## ----fitting_exp, echo=TRUE, message=FALSE, warning=FALSE---------------------
+## ----fitting_exp, message=FALSE, warning=FALSE--------------------------------
 slopes_liahovden <- flux_fitting(
   conc_df = conc_liahovden, # the output of flux_match
   f_conc = conc, # gas concentration column
@@ -57,7 +54,7 @@ flags_liahovden <- flux_quality(
 
 flux_flag_count(flags_liahovden)
 
-## ----explot, fig.width= 8, fig.height=9, message=FALSE, fig.cap="Output of `flux_plot` for fluxid 54, 95, 100 and 101."----
+## ----fig-explot, fig.width=8, fig.height=7, message=FALSE, fig.cap="Output of `flux_plot` for fluxid 54, 95, 100 and 101. With quality flags and diagnostics from `flux_quality`, the slope at $t_0$ (continuous line), the model fit (dashed line), the linear fit (dotted line), and the raw gas concentration (dots). The colours show the quality flags (green for `ok`, red for `discard` and purple for `zero` with default settings) and cuts (same colour as `discard`). The gray vertical line indicates $t_0$ (a fitting parameter when using the `exp_zhao18` model, otherwise user defined in `flux_fitting`). The g-factor is calculated as slope/linear slope, and b is the b parameter inside the exponential model. Concentration is in ppm in this example. Due to poor quality (strong peak at the start), `flux_fitting` could not provide a decent fit for fluxid 101. This was detected by `flux_quality` which flagged it as discard."----
 flags_liahovden |>
   # we show only a sample of the plots in this example
   dplyr::filter(f_fluxid %in% c(54, 95, 100, 101)) |>
@@ -71,18 +68,31 @@ flags_liahovden |>
       nrow = 2,
       ncol = 2,
       scales = "free"
-    )
+    ),
+    f_facetid = "f_fluxid"
   )
 
-## ----plot_pdf, echo=TRUE, eval=FALSE------------------------------------------
+## ----plot_pdf, eval=FALSE-----------------------------------------------------
 # flux_plot(
 #   slopes_df = flags_liahovden,
 #   f_conc = conc,
 #   f_datetime = datetime,
 #   print_plot = FALSE, # not printing the plots in the R session
-#   output = "pdfpages", # the type of output
+#   output = "longpdf", # the type of output
 #   f_plotname = "plots_liahovden" # filename for the pdf file
 # )
+
+## ----ex-plot-filter, eval=FALSE-----------------------------------------------
+# flags_lia |> # the output of flux_quality
+#   # apply here dplyr::filter on f_quality_flags, f_fluxid,
+#   # sample(f_fluxid) for random sampling of fluxid, campaigns,
+#   # dates, windspeed...
+#   # flux_plot will plot only the measurements passing the filter
+#   # and not the entire flags_lia df
+#   flux_plot(
+#     f_conc = conc,
+#     f_datetime = datetime
+#   )
 
 ## ----fits_exp_cut, message=FALSE----------------------------------------------
 fits_liahovden_60 <- conc_liahovden |>
@@ -96,11 +106,11 @@ fits_liahovden_60 <- conc_liahovden |>
 ## ----flag_exp_cut-------------------------------------------------------------
 flags_liahovden_60 <- fits_liahovden_60 |>
   flux_quality(
-    conc,
-    force_lm = 101 # we force the use of the linear model for fluxid 101
+    conc
+    # force_zero = 101 # to replace flux 101 with 0 instead of discarding
   )
 
-## ----plot_exp_cut, fig.width=8, fig.height=9, message=FALSE, fig.cap="Output of `flux_plot` for fluxid 54, 95, 100 and 101 after refitting with a 60 seconds end cut."----
+## ----fig-plot_exp_cut, fig.width=8, fig.height=9, message=FALSE, fig.cap="Output of `flux_plot` for fluxid 54, 95, 100 and 101, after quality check. Concentration is in ppm in this example."----
 flags_liahovden_60 |>
   dplyr::filter(f_fluxid %in% c(54, 95, 100, 101)) |>
   flux_plot(
@@ -128,59 +138,33 @@ fluxes_liahovden_60 <- flux_calc(
   setup_volume = 24.575, # in liters, can also be a variable
   atm_pressure = 1, # in atm, can also be a variable
   plot_area = 0.0625, # in m2, can also be a variable
-  cols_keep = c("turfID", "type"),
+  cols_keep = c("turfID", "type", "measurement_round"),
   cols_ave = c("temp_soil", "PAR")
 )
 
 ## ----fluxes-str, echo=FALSE---------------------------------------------------
-
-
 str(fluxes_liahovden_60, width = 70, strict.width = "cut")
 
-
-## ----gpp-lia------------------------------------------------------------------
-library(tidyverse)
-
-fluxes_liahovden_60 <- fluxes_liahovden_60 |>
-  mutate(
-    f_fluxid = as.integer(f_fluxid),
-    pairID = case_when(
-      type == "NEE" ~ f_fluxid,
-      type == "ER" ~ f_fluxid - 1
-    ),
-    f_fluxid = as_factor(f_fluxid),
-    pairID = as_factor(pairID)
-  )
-
-gpp_liahovden_60 <- flux_gpp(
+## ----gpp-lia, warning=FALSE---------------------------------------------------
+gpp_liahovden_60 <- flux_diff(
   fluxes_df = fluxes_liahovden_60,
   type_col = type, # the column specifying the type of measurement
-  f_datetime = datetime,
-  id_cols = c("pairID", "turfID"),
-  cols_keep = c("temp_soil_ave", "PAR_ave"), # or "none" (default) or "all"
-  nee_arg = "NEE", # default value
-  er_arg = "ER" # default value
+  id_cols = c("measurement_round", "turfID"),
+  cols_keep = c("temp_soil_ave", "PAR_ave", "datetime"), # or "none" or "all"
+  type_a = "NEE", # we want the difference between NEE
+  type_b = "ER", # and ER
+  diff_name = "GPP" # the name of the calculated flux
 )
 
 ## ----gpp-str, echo=FALSE------------------------------------------------------
-
-
 str(gpp_liahovden_60, width = 70, strict.width = "cut")
 
+## ----mg-transform-------------------------------------------------------------
+gpp_liahovden_60 <- gpp_liahovden_60 |>
+  dplyr::mutate(
+    flux_mg = f_flux * 0.0440095
+  )
 
-## ----24h_fluxes, fig.width=8, fig.height=9, warning=FALSE, message=FALSE, echo=FALSE, eval=FALSE----
-# 
-# 
-# library(ggplot2)
-# fluxes_exp_liahovden_60_gep |>
-#   ggplot(aes(x = datetime, y = f_flux)) +
-#   geom_point() +
-#   geom_smooth() +
-#   labs(
-#     title = "Net Ecosystem Exchange at Upper Site (Liahovden) during 24 hour",
-#     x = "Datetime",
-#     y = bquote(~ CO[2] ~ "flux [mmol/" * m^2 * "/h]")
-#   ) +
-#   theme(legend.position = "bottom") +
-#   facet_grid(type ~ ., scales = "free")
+## ----gpp-str2, echo=FALSE-----------------------------------------------------
+str(gpp_liahovden_60, width = 70, strict.width = "cut")